fixed UI bugs

+++ Screens/HomeScreen.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef, Component } from 'react';
+import React, { useEffect, useState, useRef, Component } from "react";
 import {
   View,
   Text,
@@ -9,67 +9,81 @@ import {
   Modal,
   Image,
   useWindowDimensions,
-} from 'react-native';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import { Button } from '@rneui/themed';
-import { WebView } from 'react-native-webview';
-import App from "../App"
+} from "react-native";
+import { MaterialCommunityIcons } from "@expo/vector-icons";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import { Button } from "@rneui/themed";
+import { WebView } from "react-native-webview";
 
 const HomeScreen = ({ navigation }) => {
-  const titleAnimation = new Animated.Value(0);
-  const welcomeAnimation = new Animated.Value(0);
-  const [modalVisible1, setmodalVisible1] = useState(false);
-  const [modalVisible2, setmodalVisible2] = useState(false);
+  const titleAnimation = useRef(new Animated.Value(0)).current;
+  const welcomeAnimation = useRef(new Animated.Value(0)).current;
+  const [modalVisible1, setModalVisible1] = useState(false);
+  const [modalVisible2, setModalVisible2] = useState(false);
+  const [isAnimating, setIsAnimating] = useState(true);
   const { width } = useWindowDimensions();
+
   const handlePress = () => {
-    setmodalVisible1(true);
+    if (!isAnimating) {
+      setModalVisible1(true);
+    }
   };
+
   const restartApp = () => {
-    AsyncStorage.clear();
-   
+    if (!isAnimating) {
+      AsyncStorage.clear();
+      navigation.reset({
+        index: 0,
+        routes: [{ name: "Home" }],
+      });
+    }
   };
+
   const closeModal1 = () => {
-    setmodalVisible1(false);
+    setModalVisible1(false);
   };
-  // For warning
+
   const handlePress1 = () => {
-    setmodalVisible2(true);
+    if (!isAnimating) {
+      setModalVisible2(true);
+    }
   };
 
   const closeModal2 = () => {
-    setmodalVisible2(false);
+    setModalVisible2(false);
   };
+
   const cards = [
     {
-      title: 'Read Paper',
-      icon: 'book-open',
-      iconColor: '#00bfa6',
-      onPress: () => handlePress(),
+      title: "Read Paper",
+      icon: "book-open",
+      iconColor: "#00bfa6",
+      onPress: handlePress,
     },
     {
-      title: 'Predictions',
-      icon: 'chart-line',
-      iconColor: '#0b7d73',
-      onPress: () => navigation.navigate('barley'),
+      title: "Predictions",
+      icon: "chart-line",
+      iconColor: "#0b7d73",
+      onPress: () => !isAnimating && navigation.navigate("barley"),
     },
     {
-      title: 'View Reports',
-      icon: 'archive-eye',
-      iconColor: '#178731',
-      onPress: () => navigation.navigate('archive'),
+      title: "View Reports",
+      icon: "archive-eye",
+      iconColor: "#178731",
+      onPress: () => !isAnimating && navigation.navigate("archive"),
     },
     {
-      title: 'Reset App',
-      icon: 'refresh',
-      iconColor: '#e6b517',
-      onPress: () => handlePress1(),
+      title: "Reset App",
+      icon: "refresh",
+      iconColor: "#e6b517",
+      onPress: handlePress1,
     },
   ];
 
-  const cardAnimations = cards.map(() => new Animated.Value(0));
+  const cardAnimations = cards.map(() => useRef(new Animated.Value(0)).current);
 
   const animateText = () => {
+    setIsAnimating(true);
     Animated.timing(titleAnimation, {
       toValue: 1,
       duration: 300,
@@ -80,14 +94,16 @@ const HomeScreen = ({ navigation }) => {
         duration: 300,
         useNativeDriver: true,
       }).start(() => {
-        cardAnimations.forEach((animation, index) => {
-          Animated.timing(animation, {
-            toValue: 1,
-            duration: 300,
-            useNativeDriver: true,
-            delay: index * 500,
-          }).start();
-        });
+        Animated.stagger(
+          500,
+          cardAnimations.map((animation) =>
+            Animated.timing(animation, {
+              toValue: 1,
+              duration: 300,
+              useNativeDriver: true,
+            })
+          )
+        ).start(() => setIsAnimating(false));
       });
     });
   };
@@ -102,21 +118,22 @@ const HomeScreen = ({ navigation }) => {
         visible={modalVisible1}
         onRequestClose={closeModal1}
         animationType="slide"
-        transparent={true}>
+        transparent={true}
+      >
         <SafeAreaView style={styles.modalContainer}>
           <View style={styles.webviewContainer}>
             <WebView
               style={styles.container}
               source={{
-                uri: 'https://www.journalresearchhs.org/_files/ugd/ebf144_ac0d6e5d6b61481d8a54a6c1377a8a84.pdf',
+                uri: "https://www.journalresearchhs.org/_files/ugd/ebf144_ac0d6e5d6b61481d8a54a6c1377a8a84.pdf",
               }}
             />
             <TouchableOpacity onPress={closeModal1} style={styles.closeButton}>
               <MaterialCommunityIcons
-                name={'window-close'}
+                name={"window-close"}
                 size={35}
-                color={'black'}
-                style={{ marginLeft: '45%' }}
+                color={"black"}
+                style={{ marginLeft: "45%" }}
               />
             </TouchableOpacity>
           </View>
@@ -126,78 +143,82 @@ const HomeScreen = ({ navigation }) => {
         visible={modalVisible2}
         onRequestClose={closeModal2}
         animationType="slide"
-        transparent={true}>
+        transparent={true}
+      >
         <SafeAreaView style={styles.modalContainer}>
           <View style={styles.warningContainer}>
             <View style={{ flex: 0.1 }}>
               <Text style={styles.titleTextForModal}>Reset Croptimization</Text>
             </View>
             <Image
-              source={require('../assets/reset.png')}
+              source={require("../assets/reset.png")}
               style={styles.image}
             />
             <View
               style={{
                 flex: 0.7,
-                width: '90%',
-                alignSelf: 'center',
-                paddingTop: '5%',
-              }}>
+                width: "90%",
+                alignSelf: "center",
+                paddingTop: "5%",
+              }}
+            >
               <Text style={styles.warningTitleTextForModal}>
                 Are you sure you want to reset the app?
               </Text>
               <Text style={styles.warningTextForModal}>
-                You will no longer be able to access{' '}
-                <Text style={{ fontWeight: 'bold', color: '#90caf9' }}>
+                You will no longer be able to access{" "}
+                <Text style={{ fontWeight: "bold", color: "#90caf9" }}>
                   app statistics, reports, and data
                 </Text>
-                . This action will permanently erase all recorded{' '}
-                <Text style={{ fontWeight: 'bold', color: '#ffcc80' }}>
+                . This action will permanently erase all recorded{" "}
+                <Text style={{ fontWeight: "bold", color: "#ffcc80" }}>
                   information, including progress, reports, and saved settings
                 </Text>
                 . Once reset, the data cannot be recovered, and the app will
-                return to its{' '}
-                <Text style={{ fontWeight: 'bold', color: '#90caf9' }}>
+                return to its{" "}
+                <Text style={{ fontWeight: "bold", color: "#90caf9" }}>
                   default state
                 </Text>
-                . Please confirm only if you’re{' '}
-                <Text style={{ fontWeight: 'bold', color: '#ffcc80' }}>
+                . Please confirm only if you’re{" "}
+                <Text style={{ fontWeight: "bold", color: "#ffcc80" }}>
                   certain
                 </Text>
-                , as this action cannot be undone. Click the{' '}
+                , as this action cannot be undone. Click the{" "}
                 <Text
                   style={{
-                    fontWeight: 'bold',
-                    color: '#ffab91',
+                    fontWeight: "bold",
+                    color: "#ffab91",
                     fontSize: 20,
-                  }}>
+                  }}
+                >
                   X
-                </Text>{' '}
+                </Text>{" "}
                 to cancel this action.
               </Text>
               <Button
                 title="Reset App"
-                buttonStyle={{ backgroundColor: '#E83A3A' }}
+                buttonStyle={{ backgroundColor: "#E83A3A" }}
                 containerStyle={{
                   height: 40,
-                  width: '95%',
-                  alignSelf: 'center',
-                  marginTop: '10%',
+                  width: "95%",
+                  alignSelf: "center",
+                  marginTop: "10%",
                   borderRadius: 10,
                 }}
-                titleStyle={{ color: 'white' }}
+                titleStyle={{ color: "white" }}
                 onPress={restartApp}
               />
             </View>
             <View style={{ flex: 0.1 }}>
               <TouchableOpacity
                 onPress={closeModal2}
-                style={styles.closeButton}>
+                style={styles.closeButton}
+              >
                 <MaterialCommunityIcons
-                  name={'window-close'}
+                  name={"window-close"}
                   size={35}
-                  color={'black'}
-                  style={{ marginLeft: '45%' }}
+                  color={"black"}
+                  style={{ marginLeft: "45%" }}
                 />
               </TouchableOpacity>
             </View>
@@ -225,9 +246,14 @@ const HomeScreen = ({ navigation }) => {
                 },
               ],
               marginBottom: 20,
-              width: '50%', // Updated for two cards per row
-            }}>
-            <TouchableOpacity style={styles.card} onPress={card.onPress}>
+              width: "50%",
+            }}
+          >
+            <TouchableOpacity
+              style={[styles.card, isAnimating && { opacity: 0.5 }]}
+              onPress={card.onPress}
+              disabled={isAnimating}
+            >
               <MaterialCommunityIcons
                 name={card.icon}
                 size={50}
@@ -247,108 +273,108 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     padding: 10,
-    backgroundColor: '#f7f7f7',
+    backgroundColor: "#f7f7f7",
   },
   image: {
     flex: 0.4,
     borderRadius: 10000,
-    alignSelf: 'center',
-    width: '100%',
-    resizeMode: 'contain',
+    alignSelf: "center",
+    width: "100%",
+    resizeMode: "contain",
   },
   titleText: {
     fontSize: 20,
-    fontWeight: 'bold',
-    marginBottom: '20%',
-    textAlign: 'center',
-    width: '100%',
+    fontWeight: "bold",
+    marginBottom: "20%",
+    textAlign: "center",
+    width: "100%",
   },
   titleTextForModal: {
     fontSize: 30,
-    fontWeight: 'bold',
-    marginBottom: '5%',
-    textAlign: 'center',
-    width: '100%',
-    color: '#ff6961',
+    fontWeight: "bold",
+    marginBottom: "5%",
+    textAlign: "center",
+    width: "100%",
+    color: "#ff6961",
   },
   welcomeText: {
     fontSize: 30,
-    fontWeight: 'bold',
+    fontWeight: "bold",
     marginBottom: 20,
-    textAlign: 'left',
-    width: '100%',
+    textAlign: "left",
+    width: "100%",
     marginLeft: 10, // Gap on the left
   },
   cardsContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-    width: '100%',
+    flexDirection: "row",
+    flexWrap: "wrap",
+    justifyContent: "space-between",
+    width: "100%",
   },
   card: {
-    backgroundColor: '#fff',
+    backgroundColor: "#fff",
     borderRadius: 10,
     padding: 20,
-    alignItems: 'center',
+    alignItems: "center",
     elevation: 5,
-    shadowColor: '#000',
+    shadowColor: "#000",
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.1,
     shadowRadius: 5,
     borderWidth: 1,
-    borderColor: '#e0e0e0',
+    borderColor: "#e0e0e0",
     height: 150,
     marginHorizontal: 15,
     marginVertical: 10,
   },
   cardText: {
-    color: '#333',
+    color: "#333",
     fontSize: 18,
-    fontWeight: 'bold',
-    textAlign: 'center',
+    fontWeight: "bold",
+    textAlign: "center",
     marginTop: 10,
   },
   warningTextForModal: {
-    color: 'gray',
+    color: "gray",
     fontSize: 17,
-    textAlign: 'left',
+    textAlign: "left",
   },
   warningTitleTextForModal: {
-    textAlign: 'left',
-    color: '#121212',
+    textAlign: "left",
+    color: "#121212",
     fontSize: 25,
-    fontWeight: 'bold',
-    textAlign: 'left',
-    marginBottom: '5%',
+    fontWeight: "bold",
+    textAlign: "left",
+    marginBottom: "5%",
   },
   icon: {
     marginBottom: 10,
   },
   modalContainer: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'white',
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor: "white",
   },
   webviewContainer: {
-    width: '100%',
-    height: '100%',
+    width: "100%",
+    height: "100%",
     borderRadius: 30,
-    overflow: 'hidden',
-    justifyContent: 'center',
+    overflow: "hidden",
+    justifyContent: "center",
   },
   warningContainer: {
-    width: '100%',
-    height: '100%',
+    width: "100%",
+    height: "100%",
     borderRadius: 30,
-    overflow: 'hidden',
-    justifyContent: 'center',
+    overflow: "hidden",
+    justifyContent: "center",
   },
   closeButton: {
     borderRadius: 90,
-    width: '100%',
-    alignSelf: 'right',
-    backgroundColor: 'white',
+    width: "100%",
+    alignSelf: "right",
+    backgroundColor: "white",
   },
 });
